requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coverage==7.2.2
cron-descriptor==1.2.35
defusedxml==0.7.1
Django==4.1.7
django-celery-beat==2.5.0
django-ckeditor==6.5.1
django-debug-toolbar==4.0.0
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.0.0
django-timezone-field==5.0
flake8==6.0.0
httpie==3.2.1
idna==3.4
kombu==5.2.4
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
